+++ src/direction/diagonal_direction.rs
@@ -74,7 +74,7 @@ pub enum DiagonalDirection {
     /// |orientation |radians|degrees|
     /// |------------|-------|-------|
     /// | Flat Top   |  π/3  |  60   |   
-    /// | Flat Top   |  π/6  |  30   |   
+    /// | Pointy Top |  π/6  |  30   |
     ///
     /// ```txt
     ///            x Axis
@@ -383,7 +383,7 @@ impl DiagonalDirection {
     ///
     /// See [`Self::angle_flat`] for *flat* hexagons
     pub fn angle_pointy(self) -> f32 {
-        self.angle_flat() - DIRECTION_ANGLE_OFFSET
+        self.angle_flat() - DIRECTION_ANGLE_OFFSET_RAD
     }
 
     #[inline]
@@ -409,6 +409,138 @@ impl DiagonalDirection {
     /// Returns the angle in radians of the given direction in the given `orientation`
     pub fn angle(self, orientation: &HexOrientation) -> f32 {
         self.angle_pointy() - orientation.angle_offset
+    }
+
+    #[inline]
+    #[must_use]
+    /// Returns the angle in degrees of the given direction according to its `orientation`
+    ///
+    /// See [`Self::angle`] for radians angles
+    pub fn angle_degrees(self, orientation: &HexOrientation) -> f32 {
+        match orientation {
+            HexOrientation::Pointy => self.angle_pointy_degrees(),
+            HexOrientation::Flat => self.angle_flat_degrees(),
+        }
+    }
+
+    #[must_use]
+    /// Returns the direction from the given `angle` in degrees
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use hexx::*;
+    ///
+    /// let direction = DiagonalDirection::from_flat_angle_degrees(15.0);
+    /// assert_eq!(direction, DiagonalDirection::Right);
+    /// ```
+    pub fn from_flat_angle_degrees(angle: f32) -> Self {
+        Self::from_pointy_angle_degrees(angle - DIRECTION_ANGLE_OFFSET_DEGREES)
+    }
+
+    #[must_use]
+    #[allow(clippy::cast_possible_truncation)]
+    /// Returns the direction from the given `angle` in degrees
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use hexx::*;
+    ///
+    /// let direction = DiagonalDirection::from_pointy_angle_degrees(15.0);
+    /// assert_eq!(direction, DiagonalDirection::TopRight);
+    /// ```
+    pub fn from_pointy_angle_degrees(angle: f32) -> Self {
+        let angle = angle.rem_euclid(360.0);
+        let sector = (angle / DIRECTION_ANGLE_DEGREES).trunc() as i32;
+        println!("{angle} - {sector}");
+        match sector {
+            0 => Self::TopRight,
+            1 => Self::TopLeft,
+            2 => Self::Left,
+            3 => Self::BottomLeft,
+            4 => Self::BottomRight,
+            _ => Self::Right,
+        }
+    }
+
+    #[must_use]
+    /// Returns the direction from the given `angle` in radians
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use hexx::*;
+    ///
+    /// let direction = DiagonalDirection::from_flat_angle(0.26);
+    /// assert_eq!(direction, DiagonalDirection::Right);
+    /// ```
+    pub fn from_flat_angle(angle: f32) -> Self {
+        Self::from_pointy_angle(angle - DIRECTION_ANGLE_OFFSET_RAD)
+    }
+
+    #[must_use]
+    #[allow(clippy::cast_possible_truncation)]
+    /// Returns the direction from the given `angle` in radians
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use hexx::*;
+    ///
+    /// let direction = DiagonalDirection::from_pointy_angle(0.26);
+    /// assert_eq!(direction, DiagonalDirection::TopRight);
+    /// ```
+    pub fn from_pointy_angle(angle: f32) -> Self {
+        let angle = angle.rem_euclid(PI_2);
+        let sector = (angle / DIRECTION_ANGLE_RAD) as i32;
+        println!("{angle} - {sector}");
+        match sector {
+            0 => Self::TopRight,
+            1 => Self::TopLeft,
+            2 => Self::Left,
+            3 => Self::BottomLeft,
+            4 => Self::BottomRight,
+            _ => Self::Right,
+        }
+    }
+
+    #[must_use]
+    /// Returns the direction from the given `angle` in degrees according the `orientation`
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use hexx::*;
+    ///
+    /// let angle = 15.0;
+    /// assert_eq!(DiagonalDirection::from_angle_degrees(angle, &HexOrientation::Flat), DiagonalDirection::Right);
+    /// assert_eq!(DiagonalDirection::from_angle_degrees(angle, &HexOrientation::Pointy), DiagonalDirection::TopRight);
+    /// ```
+    pub fn from_angle_degrees(angle: f32, orientation: &HexOrientation) -> Self {
+        match orientation {
+            HexOrientation::Pointy => Self::from_pointy_angle_degrees(angle),
+            HexOrientation::Flat => Self::from_flat_angle_degrees(angle),
+        }
+    }
+
+    #[must_use]
+    /// Returns the direction from the given `angle` in radians according the `orientation`
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use hexx::*;
+    ///
+    /// let angle = 0.26;
+    /// assert_eq!(DiagonalDirection::from_angle(angle, &HexOrientation::Flat), DiagonalDirection::Right);
+    /// assert_eq!(DiagonalDirection::from_angle(angle, &HexOrientation::Pointy), DiagonalDirection::TopRight);
+    /// ```
+    pub fn from_angle(angle: f32, orientation: &HexOrientation) -> Self {
+        match orientation {
+            HexOrientation::Pointy => Self::from_pointy_angle(angle),
+            HexOrientation::Flat => Self::from_flat_angle(angle),
+        }
     }
 
     #[deprecated(since = "0.6.0", note = "Use DiagonalDirection::direction_ccw")]