+++ src/direction/hex_direction.rs
@@ -377,7 +377,7 @@ impl Direction {
     ///
     /// See [`Self::angle_pointy`] for *pointy* hexagons
     pub fn angle_flat(self) -> f32 {
-        self.angle_pointy() + DIRECTION_ANGLE_OFFSET
+        self.angle_pointy() + DIRECTION_ANGLE_OFFSET_RAD
     }
 
     #[inline]
@@ -387,6 +387,13 @@ impl Direction {
     /// See [`Self::angle_flat`] for *flat* hexagons
     pub const fn angle_pointy(self) -> f32 {
         Self::POINTY_ANGLES[self as usize]
+    }
+
+    #[inline]
+    #[must_use]
+    /// Returns the angle in radians of the given direction in the given `orientation`
+    pub fn angle(self, orientation: &HexOrientation) -> f32 {
+        self.angle_pointy() - orientation.angle_offset
     }
 
     #[inline]
@@ -409,9 +416,132 @@ impl Direction {
 
     #[inline]
     #[must_use]
-    /// Returns the angle in radians of the given direction in the given `orientation`
-    pub fn angle(self, orientation: &HexOrientation) -> f32 {
-        self.angle_pointy() - orientation.angle_offset
+    /// Returns the angle in degrees of the given direction according to its `orientation`
+    ///
+    /// See [`Self::angle`] for radians angles
+    pub fn angle_degrees(self, orientation: &HexOrientation) -> f32 {
+        match orientation {
+            HexOrientation::Pointy => self.angle_pointy_degrees(),
+            HexOrientation::Flat => self.angle_flat_degrees(),
+        }
+    }
+
+    #[must_use]
+    /// Returns the direction from the given `angle` in degrees
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use hexx::*;
+    ///
+    /// let direction = Direction::from_pointy_angle_degrees(35.0);
+    /// assert_eq!(direction, Direction::Top);
+    /// ```
+    pub fn from_pointy_angle_degrees(angle: f32) -> Self {
+        Self::from_flat_angle_degrees(angle + DIRECTION_ANGLE_OFFSET_DEGREES)
+    }
+
+    #[must_use]
+    #[allow(clippy::cast_possible_truncation)]
+    /// Returns the direction from the given `angle` in degrees
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use hexx::*;
+    ///
+    /// let direction = Direction::from_flat_angle_degrees(35.0);
+    /// assert_eq!(direction, Direction::TopRight);
+    /// ```
+    pub fn from_flat_angle_degrees(angle: f32) -> Self {
+        let angle = angle.rem_euclid(360.0);
+        let sector = (angle / DIRECTION_ANGLE_DEGREES).trunc() as i32;
+        match sector {
+            0 => Self::TopRight,
+            1 => Self::Top,
+            2 => Self::TopLeft,
+            3 => Self::BottomLeft,
+            4 => Self::Bottom,
+            _ => Self::BottomRight,
+        }
+    }
+
+    #[must_use]
+    /// Returns the direction from the given `angle` in radians
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use hexx::*;
+    ///
+    /// let direction = Direction::from_pointy_angle(0.6);
+    /// assert_eq!(direction, Direction::Top);
+    /// ```
+    pub fn from_pointy_angle(angle: f32) -> Self {
+        Self::from_flat_angle(angle + DIRECTION_ANGLE_OFFSET_RAD)
+    }
+
+    #[must_use]
+    #[allow(clippy::cast_possible_truncation)]
+    /// Returns the direction from the given `angle` in radians
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use hexx::*;
+    ///
+    /// let direction = Direction::from_flat_angle(0.6);
+    /// assert_eq!(direction, Direction::TopRight);
+    /// ```
+    pub fn from_flat_angle(angle: f32) -> Self {
+        let angle = angle.rem_euclid(PI_2);
+        let sector = (angle / DIRECTION_ANGLE_RAD) as i32;
+        match sector {
+            0 => Self::TopRight,
+            1 => Self::Top,
+            2 => Self::TopLeft,
+            3 => Self::BottomLeft,
+            4 => Self::Bottom,
+            _ => Self::BottomRight,
+        }
+    }
+
+    #[must_use]
+    /// Returns the direction from the given `angle` in degrees according the `orientation`
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use hexx::*;
+    ///
+    /// let angle = 35.0;
+    /// assert_eq!(Direction::from_angle_degrees(angle, &HexOrientation::Flat), Direction::TopRight);
+    /// assert_eq!(Direction::from_angle_degrees(angle, &HexOrientation::Pointy), Direction::Top);
+    /// ```
+    pub fn from_angle_degrees(angle: f32, orientation: &HexOrientation) -> Self {
+        match orientation {
+            HexOrientation::Pointy => Self::from_pointy_angle_degrees(angle),
+            HexOrientation::Flat => Self::from_flat_angle_degrees(angle),
+        }
+    }
+
+    #[must_use]
+    /// Returns the direction from the given `angle` in radians according the `orientation`
+    ///
+    /// # Example
+    ///
+    /// ```rust
+    /// # use hexx::*;
+    ///
+    /// let angle = 0.6;
+    /// assert_eq!(Direction::from_angle(angle, &HexOrientation::Flat), Direction::TopRight);
+    /// assert_eq!(Direction::from_angle(angle, &HexOrientation::Pointy), Direction::Top);
+    /// ```
+    pub fn from_angle(angle: f32, orientation: &HexOrientation) -> Self {
+        match orientation {
+            HexOrientation::Pointy => Self::from_pointy_angle(angle),
+            HexOrientation::Flat => Self::from_flat_angle(angle),
+        }
     }
 
     #[deprecated(since = "0.6.0", note = "Use Direction::diagonal_ccw")]